refactor: move utils-types to string-mappers file
